third_party/WebKit/LayoutTests/battery-status/multiple-promises-expected.txt

Test multiple promise resolution.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS promise1 === promise2 is true
first resolution
PASS batteryInfo is defined.
PASS mockBatteryInfo is defined.
PASS batteryInfo.charging is mockBatteryInfo.charging
PASS batteryInfo.chargingTime is mockBatteryInfo.chargingTime
PASS batteryInfo.dischargingTime is mockBatteryInfo.dischargingTime
PASS batteryInfo.level is mockBatteryInfo.level
second resolution
PASS batteryInfo is defined.
PASS mockBatteryInfo is defined.
PASS batteryInfo.charging is mockBatteryInfo.charging
PASS batteryInfo.chargingTime is mockBatteryInfo.chargingTime
PASS batteryInfo.dischargingTime is mockBatteryInfo.dischargingTime
PASS batteryInfo.level is mockBatteryInfo.level
PASS successfullyParsed is true

TEST COMPLETE
